+++ src/components/modal.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import getPassThroughProperties from '../utilities/component/get-pass-through-properties';
 import pureRenderShouldComponentUpdate from '../utilities/pure-render-should-component-update';
+import DomEventManager from '../utilities/dom/dom-event-manager';
+import DomDimensions from '../utilities/dom/dom-dimensions';
 import getNextId from '../utilities/get-next-id';
 
 import AppendBodyComponent from './append-body-component';
@@ -11,7 +13,11 @@ class Modal extends AppendBodyComponent {
     super(props);
 
     this.uniqueId = getNextId();
+    this.domEventManager = new DomEventManager();
+
     this.setAppendElementId(this.uniqueId);
+
+    this.setDynamicHeights = this.setDynamicHeights.bind(this);
   }
 
   componentDidMount() {
@@ -19,25 +25,33 @@ class Modal extends AppendBodyComponent {
       document.querySelector('body').classList.add('modal-open');
     }
 
-    this.updateSelf();
+    if (this.props.hasDynamicHeight) {
+      this.domEventManager.add(window, 'resize', this.setDynamicHeights);
+      this.domEventManager.add(window, 'orientationchange', this.setDynamicHeights);
+    }
+
+    this.updateSelf(true);
   }
 
   componentDidUpdate(oldProps) {
+    let hideInitially = false;
     //NOTE: need to make sure when closing the modal, the scroll position is reset to the top incase it is opened again
     if (!this.props.isActive && oldProps.isActive) {
       this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"]`).scrollTop = 0;
+      this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"] .modal__content`).scrollTop = 0;
     }
 
     //NOTE we should only change the body call if the isActive has change incase there are multiple possible modals on the same page
     if (this.props.isActive !== oldProps.isActive) {
       if (this.props.isActive) {
+        hideInitially = true;
         document.querySelector('body').classList.add('modal-open');
       } else if (!this.props.isActive) {
         document.querySelector('body').classList.remove('modal-open');
       }
     }
 
-    this.updateSelf();
+    this.updateSelf(hideInitially);
   }
 
   componentWillUnmount() {
@@ -45,11 +59,34 @@ class Modal extends AppendBodyComponent {
       document.querySelector('body').classList.remove('modal-open');
     }
 
+    this.domEventManager.clear();
+
     this.removeAppendElement();
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     return pureRenderShouldComponentUpdate(this.props, nextProps, this.state, nextState);
+  }
+
+  setDynamicHeights() {
+    const modalElement = this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"] .modal`);
+    const headerElement = this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"] .modal__header`);
+    const contentElement = this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"] .modal__content`);
+    const footerElement = this.appendElementContainer.querySelector(`.modal__wrapper[data-modal-id="${this.uniqueId}"] .modal__footer`);
+
+    if (this.props.isActive) {
+      const bodyDimensions = new DomDimensions(document.body);
+      const modalDimensions = new DomDimensions(modalElement);
+      const headerDimensions = new DomDimensions(headerElement);
+      const footerDimensions = new DomDimensions(footerElement);
+
+      const modalHeight = bodyDimensions.dimensions.height - (modalDimensions.dimensions.top * 2);
+      const contentHeight = modalHeight - headerDimensions.dimensions.height - footerDimensions.dimensions.height;
+
+      contentElement.style.maxHeight = `${contentHeight}px`;
+      modalElement.style.maxHeight = `${modalHeight}px`;
+      modalElement.classList.remove('u-invisible');
+    }
   }
 
   getCssClasses() {
@@ -63,10 +100,14 @@ class Modal extends AppendBodyComponent {
       cssClasses.push('is-active');
     }
 
+    if (this.props.hasDynamicHeight) {
+      cssClasses.push('has-dynamic-height');
+    }
+
     return cssClasses;
   }
 
-  updateSelf() {
+  updateSelf(hideInitially) {
     const styles = {};
 
     if (this.props.isActive) {
@@ -81,18 +122,29 @@ class Modal extends AppendBodyComponent {
       );
     }
 
+    let modalClassName = 'modal';
+
+    if (this.props.hasDynamicHeight && hideInitially) {
+      modalClassName += ' u-invisible';
+    }
+
     this.updateAppendElement(
       <div
         data-modal-id={this.uniqueId}
         className={this.getCssClasses().join(' ')}
         {...getPassThroughProperties(this.props, 'className', 'isActive')}
       >
-        <div className="modal">
+        <div className={modalClassName}>
           {this.props.children}
         </div>
         {overlayNode}
       </div>
     );
+
+    //NOTE: this allows all styling to be applied before we calculate the dynamic heights
+    if (this.props.hasDynamicHeight) {
+      setTimeout(this.setDynamicHeights, 100);
+    }
   }
 
   render() {
@@ -104,12 +156,14 @@ Modal.displayName = 'Modal';
 
 Modal.propTypes = {
   className: React.PropTypes.string,
-  overlayDisabled: React.PropTypes.bool
+  overlayDisabled: React.PropTypes.bool,
+  hasDynamicHeight: React.PropTypes.bool
 };
 
 Modal.defaultProps = {
   className: null,
-  overlayDisabled: false
+  overlayDisabled: false,
+  hasDynamicHeight: false
 };
 
 export default Modal;