+++ web/app/pages/style-guide/wizard.page.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import Wizard from '../../../../src/components/wizard';
+import Button from '../../../../src/components/button';
 
 const codeContent = 'p {\n\tcolor: red;\n}';
 
@@ -13,11 +14,18 @@ class WizardPage extends React.Component {
     };
 
     this.closeHandler = this.closeHandler.bind(this);
+    this.openHandler = this.openHandler.bind(this);
   }
 
   closeHandler() {
     this.setState({
       wizardIsActive: false
+    });
+  }
+
+  openHandler() {
+    this.setState({
+      wizardIsActive: true
     });
   }
 
@@ -27,39 +35,27 @@ class WizardPage extends React.Component {
         <h1>Wizard</h1>
         <h2>Another Element</h2>
         <h3>3 elements on a page, that blaspheme</h3>
+        <Button onClick={this.openHandler}>open wizard</Button>
         <Wizard
           isActive={this.state.wizardIsActive}
           steps={[{
             type: 'modal',
-            content: "I'm content bitch",
+            content: "I'm content man",
             nextHandler: (increaseStep) => {
               console.log('step1 done');
               increaseStep();
             }
           }, {
             type: 'modal',
-            content: "I'm content bitch 22",
+            content: "I'm content man 22",
             className: 'modal-testing',
             previousHandler: (decreaseStep) => {
               console.log('move backwards from step2');
               decreaseStep();
             }
           }, {
-            type: 'intro',
-            highlightElements: [{
-              selector: '.p-style-guide-wizard h1',
-              padding: 5
-            }, {
-              selector: '.p-style-guide-wizard h3',
-              padding: 15
-            }],
-            className: 'testing',
-            content: (
-              <div>WHAT WHAT!!!</div>
-            )
-          }, {
             type: 'modal',
-            content: "I'm content bitch 333",
+            content: "I'm content man 333",
             nextHandler: (increaseStep) => {
               console.log('wizard done');
               increaseStep();